+++ model_navigator/framework_api/package_descriptor.py
@@ -73,6 +73,7 @@ class PackageDescriptor:
     ):
         config_filter_fields = [
             "model",
+            "model_params",
             "dataloader",
             "workdir",
             "override_workdir",