+++ CHANGELOG.md
@@ -62,7 +62,8 @@ limitations under the License.
   - new: Updated navigator package format to 0.1.1
   - new: Added Model Navigator version to status file
   - new: Add atol and rtol configuration to CLI config for model
-  - new: In case of export or conversion failures prepare minimal scripts to reproduce errors
+  - new: Added experimental support for JAX models
+  - new: In case of export or conversion failures prepere minimal scripts to reproduce errors
   - fix: conversion parameters are not stored in Navigator Package for CLI execution
 
 [//]: <> (put here on external component update with short summary what change or link to changelog)